housing/requirements.txt
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
et-xmlfile==1.1.0
idna==3.4
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.2
pdfminer.six==20221105
pycparser==2.21
pypdf==3.9.0
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
tabulate==0.9.0
tzdata==2023.3
urllib3==2.0.2
